+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.10.0] - (25.09.2021)
+## Changes
+### WARNING!! - after this update needs to be configured plugin again.
+- added switch ports control ON/OFF
+
 ## [0.9.1] - (25.09.2021)
 ## Changes
 - added possibility set policy type for configured clients